# optimize byte reading